Remove use of timestamped value.
Thought it was dumb/inelegant to have a Value() object floating around
with value and timestamp .  Instead, now all messages are sent around
as json dicts.

The request enters the system as json, flows through to an endpoint
where it becomes a reply message, then flows back to the client.