Add InternalChannel
Async I/O channel that handles connections with other nodes.  Acts as
an endpoint for received internal communications - if a request is
received it is handled right away.   Not sure if this is the best way
to do things, but creating an InternalRequest object seems redundant
since it would just process the message and push a reply back.